feat: 구매 제한 필드 추가
