+++ packages/api/src/types/admin-event.ts
@@ -98,6 +98,7 @@ export interface AdminTicketDetail {
     depositorName: string;
     depositUrl: string;
   };
+  buyTicketLimit: number;
 }
 export interface AdminTicketInfoResponse extends PaginationMeta {
   content: Content[];
